+++ src/main/duke/Duke.java
@@ -1,7 +1,14 @@
-import com.sun.tools.javac.Main;
+package duke;
 
+import com.sun.tools.javac.Main;
 import java.lang.reflect.ParameterizedType;
 import java.util.Scanner;
+import duke.command.Command;
+import duke.exception.DukeException;
+import duke.parser.Parser;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
 
 public class Duke {
     private Ui ui;
