+++ src/main/duke/command/AddCommand.java
@@ -1,3 +1,8 @@
+package duke.command;
+
+import duke.task.Task;
+import duke.task.TaskList;
+
 public class AddCommand extends Command{
     private final Task task;
