+++ src/main/duke/command/Command.java
@@ -1,3 +1,7 @@
+package duke.command;
+
+import duke.task.TaskList;
+
 public abstract class Command {
 
     public abstract void execute(TaskList tasks);
