+++ src/main/duke/exception/EmptyDescriptionException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class EmptyDescriptionException extends  DukeException{
 
     public EmptyDescriptionException(){
