+++ src/main/duke/task/Deadlines.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 public class Deadlines extends Task {
